requirements.txt
torch>=2.1.0
transformers>=4.41.0
diffusers>=0.30.0
accelerate>=0.27.0
pillow>=10.0.0
numpy>=1.24.0
openai>=1.12.0
pytest>=7.4.0
